simplify

+++ src/Grampus/Util.fs
@@ -7,9 +7,9 @@ module Util =
     let BitB i =
         Microsoft.FSharp.Core.LanguagePrimitives.EnumOfValue<uint64, Bitboard>
             (i)
-    let (-!) (r : Rank) (i : int16) : Rank = r - i
-    let (+!) (r : Rank) (i : int16) : Rank = r + i
-    let (--) (f : File) (i : int16) : File = f - i
-    let (++) (f : File) (i : int16) : File = f + i
+    let (-!) (r : Rank) (i : int) : Rank = r - i
+    let (+!) (r : Rank) (i : int) : Rank = r + i
+    let (--) (f : File) (i : int) : File = f - i
+    let (++) (f : File) (i : int) : File = f + i
